+++ resources/views/note/create.blade.php
@@ -10,7 +10,7 @@
             <x-laraknife.forms.combobox position="last" name="notestatus_scope" label="Status" :options="$optionsNotestatus"
                 width2="4" />
             <x-laraknife.forms.combobox position="first" name="visibility_scope" label="Visibility" :options="$optionsVisibility" width2="4" />
-            <x-laraknife.forms.combobox position="last" name="user_id" label="User" :options="$optionsUser" width2="4" />
+            <x-laraknife.forms.combobox position="last" name="owner_id" label="Owner" :options="$optionsUser" width2="4" />
             <x-laraknife.forms.string position="alone" name="title" label="Title"
                 value="{{ $context->valueOf('title') }}" width2="10" />
             <x-laraknife.forms.text position="alone" name="body" label="Body" value="{{ $context->valueOf('body') }}"