+++ resources/views/note/index.blade.php
@@ -11,7 +11,7 @@
       <x-laraknife.forms.string position="first" name="title" label="Title" value="{{ $context->valueOf('title') }}" width2="4" />
       <x-laraknife.forms.string position="last" name="body" label="Body" value="{{ $context->valueOf('body') }}" width2="4" />
       <x-laraknife.forms.string position="first" name="text" label="Text" value="{{ $context->valueOf('text') }}" width2="4" />
-      <x-laraknife.forms.combobox position="last" name="user" label="User" :options="$optionsUser" class="lkn-autoupdate" width2="4" />
+      <x-laraknife.forms.combobox position="last" name="owner" label="Owner" :options="$optionsUser" class="lkn-autoupdate" width2="4" />
       </x-laraknife.panels.filter>
       <x-laraknife.panels.index-button buttonType="new"/>
       <x-laraknife.panels.sortable-table :context="$context" :pagination="$pagination">
@@ -22,7 +22,7 @@
             <th sortId="body">{{__('Body')}}</th>
             <th sortId="category_scope">{{__('Category')}}</th>
             <th sortId="notestatus_scope">{{__('Status')}}</th>
-            <th sortId="user_id">{{__('User')}}</th>
+            <th sortId="owner_id">{{__('Owner')}}</th>
             <th></th>
           </tr>
         </thead>
@@ -34,7 +34,7 @@
               <td>{{$note->body_short}}</td>
               <td>{{ __($note->category_scope) }}</td>
               <td>{{ __($note->notestatus_scope) }}</td>
-              <td>{{$note->user_id}}</td>
+              <td>{{$note->owner_id}}</td>
             <td><x-laraknife.icons.delete-record module="note" no="{{ $note->id }}" /></td>
         </tr>
 @endforeach