+++ resources/views/user/login.blade.php
@@ -16,7 +16,7 @@
             <x-laraknife.forms.checkbox position="last" name="autologin" label="Remain signed in" labelBelow="true" width2="4" />
             <x-laraknife.forms.set-position position="first" />
             <x-laraknife.layout.col-empty width="3" />
-            <x-laraknife.buttons.button-position position="last" name="btnLogin" label="{{ __('Login') }}" />
+            <x-laraknife.buttons.button-position position="last" name="btnLogin" label="Login" />
             <x-laraknife.layout.row-empty />
             <x-laraknife.forms.link position="alone" reference="/user-forgotten" text="{{__('Password forgotten?')}}" width1="5" width2="4" />
         </x-laraknife.panels.standard>